move utils function outside index
